Initial strategy based refactoring of stimuli generator.
