--- tt_metal/tt-llk/tests/python_tests/helpers/stimuli_generator.py
@@ -1,1937 +1,0 @@
-# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
-# SPDX-License-Identifier: Apache-2.0
-
-import math
-import warnings
-from dataclasses import dataclass
-from enum import Enum
-from typing import Callable, Dict, List, Optional, Set, Tuple, Union
-
-import torch
-
-from .bfp_format_utils import bfp4b_to_float16b
-from .format_config import MX_FORMAT_MAX_NORMAL, MX_FORMAT_MIN_MAGNITUDE, DataFormat
-from .llk_params import format_dict
-from .tile_constants import (
-    DEFAULT_TILE_C_DIM,
-    DEFAULT_TILE_R_DIM,
-    FACE_C_DIM,
-    MAX_FACE_R_DIM,
-    MAX_NUM_FACES,
-    get_tile_params,
-    validate_tile_dimensions,
-)
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Distribution kinds
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-class DistributionKind(str, Enum):
-    UNIFORM = "uniform"
-    GAUSSIAN = "gaussian"
-    SAW = "saw"
-    LOG_UNIFORM = "log_uniform"
-    CONSTANT = "constant"
-    SEQUENTIAL = "sequential"
-    RAMP = "ramp"
-    GAUSSIAN_LINSPACE = "gaussian_linspace"
-    LOG_UNIFORM_LINSPACE = "log_uniform_linspace"
-    IDENTITY = "identity"
-    FACE_IDENTITY = "face_identity"
-    CUSTOM = "custom"
-    ULP_SWEEP = "ulp_sweep"
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# StimuliSpec
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-@dataclass
-class StimuliSpec:
-    """
-    Declarative per-operand specification for LLK test stimuli generation.
-
-    Parameters
-    ----------
-    distribution : DistributionKind or callable
-        How values are sampled. Supported enum members (from DistributionKind):
-
-        "uniform"
-            Uniform random in [low, high]. For integer formats the bounds
-            are narrowed to the tightest enclosing integers via
-            torch.randint(ceil(low), floor(high) + 1), so only integers
-            that actually lie within [low, high] are generated.
-
-        "gaussian"
-            Normal distribution with *mean* and *std*. For integer formats
-            the result is rounded and clamped to the representable range.
-
-        "saw"
-            Per-face sawtooth: linearly spaced values from *low* to *high*
-            (torch.linspace), restarting every face.  For integer
-            formats a float32 linspace is rounded and clamped.
-
-        "log_uniform"
-            exp(Uniform(log(low), log(high))).  Both *low* and *high* must
-            be strictly positive.
-
-        "constant"
-            Every element is set to *value*.  Ignores *low*, *high*, *seed*.
-
-        "sequential"
-            Arithmetic progression.  By default generates 1, 2, 3, …,
-            size (legacy behavior).  Use the ``sequential()`` factory
-            with *low*, *high*, and *step* to customize.  When *high*
-            is set, positions beyond it are zero-filled.  Supports
-            *intervals* — values outside the union are zeroed.
-
-        "ramp"
-            Continuous linear sweep: deterministic, evenly spaced values
-            from *low* to *high* (torch.linspace) across the full
-            tensor (short-circuits the face loop), producing one smooth
-            ramp instead of a per-face sawtooth.  Useful for plotting
-            function shapes.
-
-        "gaussian_linspace"
-            Deterministic sweep through the Gaussian domain via the
-            inverse CDF (percent-point function).  Produces ordered values
-            concentrated around *mean* and spreading into the tails at
-            *std* scale.  No randomness involved — the output is fully
-            determined by *mean*, *std*, and the element count.
-
-        "log_uniform_linspace"
-            Deterministic, logarithmically spaced values from *low* to
-            *high*.  Both bounds must be strictly positive.  Equivalent to
-            torch.logspace in natural-log base.
-
-        "identity"
-            Identity matrix pattern: *value* on the diagonal, zero
-            elsewhere.  Requires 2-D input_dimensions.  For
-            rectangular matrices the diagonal has min(rows, cols)
-            entries.  Bypasses the face loop (tensor-level operation).
-            Ignores *low*, *high*, *seed*.
-
-        "face_identity"
-            Per-face identity block: each face is treated as a
-            (face_r_dim, FACE_C_DIM) matrix with *value* on the
-            diagonal and zero elsewhere.  The diagonal length is
-            min(face_r_dim, FACE_C_DIM).  Participates in the
-            normal face loop, so it works naturally with face_specs
-            and masked_faces.  Ignores *low*, *high*, *seed*.
-
-        "custom"
-            Explicit per-face values: the elements from *values* are
-            written at the start of the flattened face, and every
-            remaining element is zero.  Values are not repeated.
-            Participates in the normal face loop, so it works naturally
-            with face_specs and masked_faces.  Ignores *low*,
-            *high*, *seed*, *value*.
-
-        "ulp_sweep"
-            Exhaustive 1-ULP sweep: enumerates every finite representable
-            value in [low, high] for the target format (sorted, deduplicated),
-            pads with zeros to fill the tensor.  Only Float16_b and
-            Float16 formats are supported.  Bypasses the face loop
-            (tensor-level operation).  Ignores *seed*, *mean*, *std*,
-            *value*, *intervals*, *face_specs*, *masked_faces*.
-
-            **Dimension auto-sizing (operand A only):** when spec_A uses
-            ULP_SWEEP, generate_stimuli ignores any caller-supplied
-            input_dimensions_A and computes a 32×(32*N) layout large
-            enough to hold all representable values.  When spec_B is
-            omitted, input_dimensions_B is mirrored from A.  If spec_B
-            is provided explicitly, input_dimensions_B is *not*
-            auto-sized — the caller must supply compatible dimensions.
-
-        callable
-            fn(size: int, dtype: torch.dtype, generator: Optional[torch.Generator]) -> torch.Tensor.
-            The *generator* argument carries the per-operand RNG state (or
-            "None" when no seed is set), enabling reproducible custom
-            distributions.  The caller is responsible for producing a 1-D
-            tensor of exactly *size* elements and returning it as the
-            requested *dtype*.
-
-    low: float
-        Lower bound for "uniform", "saw", "ramp",
-        "log_uniform", and "log_uniform_linspace".
-        Defaults to 0.0.
-    high: float
-        Upper bound for "uniform", "saw", "ramp",
-        "log_uniform", and "log_uniform_linspace".
-        Defaults to 1.0.
-    value: float
-        Fill value used by "constant" (all elements), "identity"
-        (diagonal elements), and "face_identity" (face diagonal
-        elements).  Defaults to 1.0.
-    values: list[float], optional
-        Explicit value list for "custom".  Written at the start of
-        the flattened face; remaining elements are zero.  For integer
-        formats each value is rounded and clamped to the representable
-        range.  Must not be empty when distribution="custom".
-    mean: float
-        Mean for "gaussian" and "gaussian_linspace".  Defaults to
-        0.0.
-    std: float
-        Standard deviation for "gaussian" and "gaussian_linspace".
-        Defaults to 1.0.
-    seed: int, optional
-        Seed for a per-spec torch.Generator.  "None" uses the global
-        torch RNG state.  When an external generator is supplied to
-        generate_face function the *seed* field is ignored so the caller
-        controls state across faces.
-    face_specs: list[StimuliSpec | None], optional
-        Per-face overrides. face_specs itself may be "None"
-        (no overrides at all).  Individual entries may also be "None",
-        meaning "use the outer/base spec for this face."  Face *i* is
-        generated with face_specs[i] when that entry is a
-        StimuliSpec class, a "None" entry or an index beyond the
-        list length falls back to the outer spec.
-    masked_faces: set[int], optional
-        Set of 0-based face indices whose output should be zeroed.
-        Masking is applied *after* face generation (including any
-        face_specs overrides), so it always wins.  Not compatible
-        with global short-circuit distributions ("identity",
-        "sequential", linspace variants) — raises ValueError
-        if combined.
-    intervals: list[tuple[float, float]], optional
-        Union of [low, high] ranges.  When set, *low*/*high* are ignored
-        and values are generated from the union of these intervals.
-        Supported distributions:
-
-        - **uniform** (random): each interval is selected with
-          probability proportional to its length.
-        - **log_uniform** (random): each interval is selected with
-          probability proportional to its log-space length.
-        - **integer uniform**: each interval is selected with
-          probability proportional to the number of integer points.
-        - **gaussian**: truncated Gaussian via rejection sampling —
-          only values inside the union are kept.  For integer formats
-          the samples are rounded and clamped to representable integers
-          within the intervals.
-        - **saw**: piecewise linspace — each interval gets a proportional
-          share of the total *size* elements, concatenated in order.
-        - **ramp / log_uniform_linspace**: same piecewise linspace
-          semantics as *saw*, applied at the tensor level.
-        - **sequential**: values outside the union are zeroed.
-
-        Not supported for *gaussian_linspace* (raises ValueError).
-        Ignored by *identity*, *face_identity*, *custom*.
-        Example::
-
-            StimuliSpec.uniform(intervals=[(-10.0, -1.0), (1.0, 10.0)])
-    """
-
-    distribution: Union[DistributionKind, Callable] = DistributionKind.UNIFORM
-    low: float = 0.0
-    high: float = 1.0
-    value: float = 1.0
-    values: Optional[List[float]] = None
-    mean: float = 0.0
-    std: float = 1.0
-    seed: Optional[int] = None
-    face_specs: Optional[List[Optional["StimuliSpec"]]] = None
-    masked_faces: Optional[Set[int]] = None
-    intervals: Optional[List[Tuple[float, float]]] = None
-
-    def __post_init__(self) -> None:
-        if not (
-            callable(self.distribution)
-            or isinstance(self.distribution, DistributionKind)
-        ):
-            raise TypeError(
-                f"StimuliSpec.distribution must be DistributionKind or callable, "
-                f"got {type(self.distribution).__name__!r}: {self.distribution!r}"
-            )
-
-    # ── convenience constructors ──────────────────────────────────────────────
-
-    @classmethod
-    def constant(cls, value: float = 1.0, **kwargs) -> "StimuliSpec":
-        """All elements equal to *value*."""
-        return cls(distribution=DistributionKind.CONSTANT, value=value, **kwargs)
-
-    @classmethod
-    def identity(cls, diagonal_value: float = 1.0, **kwargs) -> "StimuliSpec":
-        """Identity matrix: *diagonal_value* on the diagonal, zero elsewhere."""
-        return cls(
-            distribution=DistributionKind.IDENTITY, value=diagonal_value, **kwargs
-        )
-
-    @classmethod
-    def face_identity(cls, diagonal_value: float = 1.0, **kwargs) -> "StimuliSpec":
-        """Per-face identity block: *diagonal_value* on the face diagonal, zero elsewhere."""
-        return cls(
-            distribution=DistributionKind.FACE_IDENTITY, value=diagonal_value, **kwargs
-        )
-
-    @classmethod
-    def custom(cls, values: List[float], **kwargs) -> "StimuliSpec":
-        """Explicit values at the start of each face, zero-filled remainder."""
-        return cls(distribution=DistributionKind.CUSTOM, values=list(values), **kwargs)
-
-    @classmethod
-    def custom_faces(
-        cls,
-        face_values: Dict[int, List[float]],
-        **kwargs,
-    ) -> "StimuliSpec":
-        """
-        Build a spec that writes custom values on selected faces and leaves all
-        other faces as zeros.
-
-        Args:
-            face_values:
-                Mapping from 0-based face index to the value list for that face.
-                Each list is written at the start of the flattened face; the rest
-                of the face is zero-filled. Faces not present in the mapping are
-                entirely zeros.
-            **kwargs: Forwarded to the outer StimuliSpec (e.g. masked_faces).
-
-        Returns:
-            StimuliSpec with distribution="constant", value=0.0 as the base and a
-            face_specs list that overlays per-face custom(values=...) specs
-            where provided.
-        """
-        if not face_values:
-            raise ValueError("face_values must be a non-empty dict")
-        if any(idx < 0 for idx in face_values):
-            raise ValueError(
-                f"face_values keys must be non-negative, "
-                f"got {sorted(k for k in face_values if k < 0)}"
-            )
-        if "face_specs" in kwargs:
-            raise ValueError(
-                "Cannot combine custom_faces() with an explicit face_specs "
-                "argument — custom_faces builds face_specs internally"
-            )
-        max_idx = max(face_values)
-        face_specs: List[Optional["StimuliSpec"]] = [None] * (max_idx + 1)
-        for idx, vals in face_values.items():
-            face_specs[idx] = cls.custom(values=vals)
-        return cls(
-            distribution=DistributionKind.CONSTANT,
-            value=0.0,
-            face_specs=face_specs,
-            **kwargs,
-        )
-
-    @classmethod
-    def sequential(
-        cls,
-        low: Optional[float] = None,
-        high: Optional[float] = None,
-        step: float = 1.0,
-        **kwargs,
-    ) -> "StimuliSpec":
-        """Arithmetic progression, optionally bounded with zero-fill.
-
-        No arguments produces the legacy 1, 2, 3, … sequence.  With
-        *low*/*high*/*step*, generates low, low+step, low+2*step, …
-        up to *high* (then zero-fills the remainder).
-        """
-        kw = dict(distribution=DistributionKind.SEQUENTIAL, **kwargs)
-        if low is not None:
-            kw["low"] = low
-        if high is not None:
-            kw["high"] = high
-        kw["std"] = step
-        return cls(**kw)
-
-    @classmethod
-    def uniform(cls, low: float = 0.0, high: float = 1.0, **kwargs) -> "StimuliSpec":
-        """Uniform random in [low, high]."""
-        return cls(distribution=DistributionKind.UNIFORM, low=low, high=high, **kwargs)
-
-    @classmethod
-    def gaussian(cls, mean: float = 0.0, std: float = 1.0, **kwargs) -> "StimuliSpec":
-        """Normal distribution with *mean* and *std*."""
-        return cls(distribution=DistributionKind.GAUSSIAN, mean=mean, std=std, **kwargs)
-
-    @classmethod
-    def saw(cls, low: float = 0.0, high: float = 1.0, **kwargs) -> "StimuliSpec":
-        """Per-face sawtooth linspace from *low* to *high*, restarting every face."""
-        return cls(distribution=DistributionKind.SAW, low=low, high=high, **kwargs)
-
-    @classmethod
-    def log_uniform(
-        cls, low: float = 1e-4, high: float = 1.0, **kwargs
-    ) -> "StimuliSpec":
-        """Log-uniform in [low, high]. Both bounds must be strictly positive."""
-        if low <= 0 or high <= 0:
-            raise ValueError(
-                f"log_uniform requires strictly positive low and high, "
-                f"got low={low}, high={high}"
-            )
-        return cls(
-            distribution=DistributionKind.LOG_UNIFORM, low=low, high=high, **kwargs
-        )
-
-    @classmethod
-    def ramp(cls, low: float = 0.0, high: float = 1.0, **kwargs) -> "StimuliSpec":
-        """Continuous linear sweep from *low* to *high* across the full tensor."""
-        return cls(distribution=DistributionKind.RAMP, low=low, high=high, **kwargs)
-
-    @classmethod
-    def gaussian_linspace(
-        cls, mean: float = 0.0, std: float = 1.0, **kwargs
-    ) -> "StimuliSpec":
-        """
-        Deterministic sweep through a Gaussian: values spaced by inverting
-        the normal CDF (no randomness, fully determined by mean/std/size).
-        """
-        return cls(
-            distribution=DistributionKind.GAUSSIAN_LINSPACE,
-            mean=mean,
-            std=std,
-            **kwargs,
-        )
-
-    @classmethod
-    def log_uniform_linspace(
-        cls, low: float = 1e-4, high: float = 1.0, **kwargs
-    ) -> "StimuliSpec":
-        """Deterministic log-spaced sweep from *low* to *high* (both > 0)."""
-        if low <= 0 or high <= 0:
-            raise ValueError(
-                f"log_uniform_linspace requires strictly positive low and high, "
-                f"got low={low}, high={high}"
-            )
-        return cls(
-            distribution=DistributionKind.LOG_UNIFORM_LINSPACE,
-            low=low,
-            high=high,
-            **kwargs,
-        )
-
-    @classmethod
-    def ulp_sweep(cls, low: float, high: float, **kwargs) -> "StimuliSpec":
-        """Exhaustive 1-ULP sweep of all representable values in [low, high].
-
-        Enumerates every finite representable value for the target format,
-        sorted and deduplicated, padding with zeros to fill the tensor.
-        Only Float16_b and Float16 formats are supported.
-
-        When used as spec_A, generate_stimuli auto-sizes input_dimensions_A
-        and mirrors it to B if spec_B is omitted.  Explicit spec_B is not
-        auto-sized — the caller must supply compatible input_dimensions_B.
-        """
-        return cls(
-            distribution=DistributionKind.ULP_SWEEP, low=low, high=high, **kwargs
-        )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Tile / face count utilities
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def calculate_tile_and_face_counts(
-    input_dimensions_A: list,
-    input_dimensions_B: list,
-    face_r_dim: int,
-    num_faces: int,
-) -> tuple[int, int, int]:
-    """
-    Calculate tile counts and faces to generate based on input dimensions and face configuration.
-    Uses 32x32 tiles in the full-face case; falls back to a single-tile
-    partial-face layout when face_r_dim < MAX_FACE_R_DIM.
-
-    Args:
-        input_dimensions_A: [height, width] in elements for input A
-        input_dimensions_B: [height, width] in elements for input B
-        face_r_dim: Number of rows in a face (typically 16 for full faces)
-        num_faces: Number of faces to generate for partial face case
-
-    Returns:
-        tuple: (tile_cnt_A, tile_cnt_B, faces_to_generate)
-    """
-    assert face_r_dim == MAX_FACE_R_DIM or (
-        face_r_dim < MAX_FACE_R_DIM and face_r_dim == input_dimensions_A[0]
-    ), (
-        f"face_r_dim must be {MAX_FACE_R_DIM} (full face) or < {MAX_FACE_R_DIM} "
-        f"and equal to input_dimensions_A[0] (partial single-tile case); "
-        f"got face_r_dim={face_r_dim}, input_dimensions_A={input_dimensions_A}"
-    )
-
-    # Handle partial faces
-    if face_r_dim < MAX_FACE_R_DIM:
-        # Partial face case: generate exactly num_faces worth of data
-        tile_cnt_A, tile_cnt_B = 1, 1
-        faces_to_generate = num_faces  # Generate exactly the right number of faces
-    else:
-        # Full tile case - always use 32x32 tiles
-        tile_cnt_A = (
-            input_dimensions_A[0]
-            // DEFAULT_TILE_R_DIM
-            * input_dimensions_A[1]
-            // DEFAULT_TILE_C_DIM
-        )
-        tile_cnt_B = (
-            input_dimensions_B[0]
-            // DEFAULT_TILE_R_DIM
-            * input_dimensions_B[1]
-            // DEFAULT_TILE_C_DIM
-        )
-        faces_to_generate = MAX_NUM_FACES
-
-    return tile_cnt_A, tile_cnt_B, faces_to_generate
-
-
-def calculate_tile_and_face_counts_w_tile_dimensions(
-    input_dimensions_A: list,
-    input_dimensions_B: list,
-    face_r_dim: int,
-    num_faces: int,
-    tile_dimensions: list,
-) -> tuple[int, int, int]:
-    """
-    Calculate tile counts and faces to generate for variable tile dimensions (dense mode).
-
-    Args:
-        input_dimensions_A: [height, width] in elements for input A
-        input_dimensions_B: [height, width] in elements for input B
-        face_r_dim: Number of rows in a face (1, 2, 4, 8, or 16)
-        num_faces: Number of faces per tile (1, 2, or 4)
-        tile_dimensions: [rows, cols] for tile size
-
-    Returns:
-        tuple: (tile_cnt_A, tile_cnt_B, faces_to_generate)
-    """
-    validate_tile_dimensions(tile_dimensions)
-    tile_r_dim, tile_c_dim = tile_dimensions
-
-    # Calculate tile counts based on actual tile dimensions
-    tile_cnt_A = (input_dimensions_A[0] // tile_r_dim) * (
-        input_dimensions_A[1] // tile_c_dim
-    )
-    tile_cnt_B = (input_dimensions_B[0] // tile_r_dim) * (
-        input_dimensions_B[1] // tile_c_dim
-    )
-    # Always generate all faces to fill the tile densely
-    faces_to_generate = num_faces
-
-    return tile_cnt_A, tile_cnt_B, faces_to_generate
-
-
-def _clamp_mx_tensors(
-    srcA_tensor: torch.Tensor,
-    srcB_tensor: torch.Tensor,
-    stimuli_format_A: DataFormat,
-    stimuli_format_B: DataFormat,
-    output_format: DataFormat = None,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """
-    Clamp tensors for MX format compatibility.
-
-    Args:
-        srcA_tensor: Source A tensor
-        srcB_tensor: Source B tensor
-        stimuli_format_A: Data format for source A
-        stimuli_format_B: Data format for source B
-        output_format: Optional output format for range constraints
-
-    Returns:
-        tuple: (clamped_srcA_tensor, clamped_srcB_tensor)
-    """
-    # Clamp inputs if both are different MX formats (use more restrictive MxFp8P)
-    if stimuli_format_A.is_mx_format() and stimuli_format_B.is_mx_format():
-        if stimuli_format_A != stimuli_format_B:
-            srcA_tensor = torch.clamp(
-                srcA_tensor,
-                -MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-                MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-            )
-            srcB_tensor = torch.clamp(
-                srcB_tensor,
-                -MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-                MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-            )
-
-    # Clamp inputs based on output format to prevent excessive rounding errors
-    if output_format == DataFormat.MxFp8P:
-        srcA_tensor = torch.clamp(
-            srcA_tensor,
-            -MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-            MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-        )
-        srcB_tensor = torch.clamp(
-            srcB_tensor,
-            -MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-            MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P],
-        )
-    elif output_format == DataFormat.MxFp8R:
-        srcA_tensor = torch.clamp(
-            srcA_tensor,
-            -MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8R],
-            MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8R],
-        )
-        srcB_tensor = torch.clamp(
-            srcB_tensor,
-            -MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8R],
-            MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8R],
-        )
-
-    return srcA_tensor, srcB_tensor
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Matmul helpers
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def _mask_tile(
-    tile: torch.Tensor,
-    num_faces: int,
-    is_matrix_B: bool,
-    face_r_dim: int = MAX_FACE_R_DIM,
-) -> torch.Tensor:
-    masked = tile.clone()
-    if num_faces == 1:
-        # Keep only f0
-        masked[:MAX_FACE_R_DIM, FACE_C_DIM:] = 0  # Zero f1
-        masked[face_r_dim:, :] = 0  # Zero f2, f3 and part of f0
-    elif num_faces == 2:
-        if is_matrix_B:
-            # matrix B (In1/SrcA): keep partial f0, f2
-            if face_r_dim < MAX_FACE_R_DIM:
-                masked[face_r_dim:MAX_FACE_R_DIM, :FACE_C_DIM] = 0  # Zero part of f0
-                masked[MAX_FACE_R_DIM + face_r_dim :, :FACE_C_DIM] = (
-                    0  # Zero part of f2
-                )
-            masked[:MAX_FACE_R_DIM, FACE_C_DIM:] = 0  # Zero f1
-            masked[MAX_FACE_R_DIM:, FACE_C_DIM:] = 0  # Zero f3
-        else:
-            # matrix A (In0/SrcB): keep f0, f1
-            masked[face_r_dim:, :] = 0  # Zero part of f0 and f1
-    return masked
-
-
-def generate_face_matmul_data(
-    num_faces: int,
-    stimuli_format: DataFormat,
-    input_dimensions=[DEFAULT_TILE_R_DIM, DEFAULT_TILE_C_DIM],
-    is_matrix_A=True,  # True for matrix A (SrcB), False for matrix B (SrcA)
-    face_r_dim=MAX_FACE_R_DIM,
-) -> torch.Tensor:
-
-    # Validate num_faces
-    if num_faces not in [1, 2, MAX_NUM_FACES]:
-        raise ValueError(f"num_faces must be 1, 2, or {MAX_NUM_FACES}, got {num_faces}")
-
-    # Validate input_dimensions
-    rows, cols = input_dimensions
-    if rows % DEFAULT_TILE_R_DIM != 0 or cols % DEFAULT_TILE_C_DIM != 0:
-        raise ValueError(
-            f"Input dimensions must be multiples of {DEFAULT_TILE_R_DIM}, "
-            f"got {input_dimensions}"
-        )
-
-    rt, ct = rows // DEFAULT_TILE_R_DIM, cols // DEFAULT_TILE_C_DIM
-    dtype = format_dict[stimuli_format]
-
-    out = torch.rand(rt, ct, DEFAULT_TILE_R_DIM, DEFAULT_TILE_C_DIM, dtype=dtype)
-    mask = _mask_tile(
-        torch.ones(DEFAULT_TILE_R_DIM, DEFAULT_TILE_C_DIM, dtype=dtype),
-        num_faces,
-        not is_matrix_A,
-        face_r_dim,
-    )
-    out *= mask
-    out = out.permute(0, 2, 1, 3).reshape(rows, cols)
-
-    return out
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# L1 view helpers
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def convert_to_l1_view(
-    tilized_tensor: torch.Tensor,
-    input_dimensions: list,
-    tile_dimensions: list = None,
-) -> torch.Tensor:
-    """
-    Convert a tilized tensor to its L1 memory view by condensing data based on tile dimensions.
-
-    This function extracts only the data that corresponds to the specified tile dimensions
-    and places it at the beginning of each tile, with the remaining space zeroed out.
-    The full tile size (1024 elements) is preserved.
-
-    Tilized format: faces are stored sequentially [f0 (256), f1 (256), f2 (256), f3 (256)]
-    Within each face, data is stored row-major (16 rows × 16 cols).
-
-    Face layout in a 32×32 tile:
-    - f0: rows 0-15, cols 0-15  (top-left)
-    - f1: rows 0-15, cols 16-31 (top-right)
-    - f2: rows 16-31, cols 0-15 (bottom-left)
-    - f3: rows 16-31, cols 16-31 (bottom-right)
-
-    Examples:
-    - tile_dimensions=[32, 32]: full tile, no change [f0, f1, f2, f3]
-    - tile_dimensions=[16, 32]: top half [f0, f1, 0, 0]
-    - tile_dimensions=[32, 16]: left half [f0, f2, 0, 0]
-    - tile_dimensions=[16, 16]: top-left only [f0, 0, 0, 0]
-    - tile_dimensions=[8, 32]: first 8 rows [f0_rows0-7, f1_rows0-7, 0, ...]
-
-    Args:
-        tilized_tensor: Input tensor in tilized format (faces stored sequentially per tile)
-        input_dimensions: [rows, cols] of the full input matrix
-        tile_dimensions: [rows, cols] to keep per tile (default [32, 32])
-                        rows must be one of: 1, 2, 4, 8, 16, 32
-                        cols must be one of: 16, 32
-
-    Returns:
-        Tensor with condensed data at the beginning (face by face), zeros at the end
-    """
-    if tile_dimensions is None:
-        tile_dimensions = [32, 32]
-
-    tile_rows, tile_cols = tile_dimensions
-
-    valid_rows = {1, 2, 4, 8, 16, 32}
-    valid_cols = {16, 32}
-
-    if tile_rows not in valid_rows:
-        raise ValueError(
-            f"tile_dimensions[0] (rows) must be one of {sorted(valid_rows)}, got {tile_rows}"
-        )
-    if tile_cols not in valid_cols:
-        raise ValueError(
-            f"tile_dimensions[1] (cols) must be one of {sorted(valid_cols)}, got {tile_cols}"
-        )
-
-    rows, cols = input_dimensions
-    if rows % 32 != 0 or cols % 32 != 0:
-        raise ValueError(
-            f"Input dimensions must be multiples of 32, got {input_dimensions}"
-        )
-
-    # If using full tile dimensions, no conversion needed
-    if tile_rows == 32 and tile_cols == 32:
-        return tilized_tensor.flatten()
-
-    # Calculate number of tiles
-    tile_cnt = (rows // 32) * (cols // 32)
-    face_rows = 16
-    face_cols = 16
-
-    # Reshape to [num_tiles, 4, 16, 16] for easier face/row manipulation
-    # Face order in tilized format: [f0, f1, f2, f3]
-    tensor_by_tiles = tilized_tensor.flatten().view(tile_cnt, 4, face_rows, face_cols)
-
-    # Create output tensor with same shape, initialized to zeros
-    output = torch.zeros_like(tensor_by_tiles)
-
-    # Determine which faces to use and how many rows from each
-    # tile_rows <= 16: only top faces (f0, f1), take tile_rows from each
-    # tile_rows == 32: all faces, take all 16 rows from each
-    # tile_cols == 16: only left faces (f0, f2)
-    # tile_cols == 32: both left and right faces
-    use_bottom_faces = tile_rows == 32
-    use_right_faces = tile_cols == 32
-    rows_per_face = tile_rows if tile_rows <= 16 else 16
-
-    # Extract data face by face (not interleaved)
-    for tile_idx in range(tile_cnt):
-        out_flat = []
-
-        # f0: always used - extract rows_per_face rows
-        for row in range(rows_per_face):
-            out_flat.extend(tensor_by_tiles[tile_idx, 0, row, :].tolist())
-
-        # f1: used if tile_cols == 32 - extract rows_per_face rows
-        if use_right_faces:
-            for row in range(rows_per_face):
-                out_flat.extend(tensor_by_tiles[tile_idx, 1, row, :].tolist())
-
-        # f2: used if tile_rows == 32 - extract all 16 rows
-        if use_bottom_faces:
-            for row in range(16):
-                out_flat.extend(tensor_by_tiles[tile_idx, 2, row, :].tolist())
-
-        # f3: used if tile_rows == 32 and tile_cols == 32 - extract all 16 rows
-        if use_bottom_faces and use_right_faces:
-            for row in range(16):
-                out_flat.extend(tensor_by_tiles[tile_idx, 3, row, :].tolist())
-
-        # Place condensed data at the beginning of the tile
-        out_flat_tensor = torch.tensor(out_flat, dtype=tilized_tensor.dtype)
-        output_flat = output[tile_idx].flatten()
-        output_flat[: len(out_flat)] = out_flat_tensor
-        output[tile_idx] = output_flat.view(4, face_rows, face_cols)
-
-    # Flatten and return
-    return output.flatten()
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Private helpers
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-# Distributions that generate one global deterministic sweep across the tensor
-# (linspace-style) and therefore bypass per-face generation.
-_LINSPACE_DISTRIBUTIONS = frozenset(
-    {
-        DistributionKind.RAMP,
-        DistributionKind.GAUSSIAN_LINSPACE,
-        DistributionKind.LOG_UNIFORM_LINSPACE,
-    }
-)
-
-# Tiny offset so we never hit exact 0 or 1 in the Gaussian CDF,
-# which would otherwise produce +/- infinity.
-_GAUSSIAN_LINSPACE_EPS = 1e-6
-
-
-def _generate_linspace_tensor(
-    spec: StimuliSpec, size: int, dtype: torch.dtype
-) -> torch.Tensor:
-    """Generate a deterministic linspace-style tensor (ramp/gaussian_linspace/log_uniform_linspace).
-
-    Args:
-        spec: Stimuli specification with distribution type and bounds.
-        size: Number of elements to generate.
-        dtype: Target torch dtype (computation uses float32, cast at end).
-
-    Returns:
-        1-D tensor of *size* elements in the requested dtype.
-    """
-    dist = spec.distribution
-
-    if dist == DistributionKind.RAMP:
-        if spec.intervals:
-            counts = _split_size_across_intervals(spec.intervals, size)
-            segments = []
-            for (lo, hi), n in zip(spec.intervals, counts):
-                if n > 0:
-                    segments.append(torch.linspace(lo, hi, n, dtype=torch.float32))
-            return torch.cat(segments).to(dtype=dtype)
-        return torch.linspace(spec.low, spec.high, size, dtype=torch.float32).to(
-            dtype=dtype
-        )
-
-    if dist == DistributionKind.GAUSSIAN_LINSPACE:
-        if spec.intervals:
-            raise ValueError("intervals not supported for gaussian_linspace yet")
-        p = torch.linspace(_GAUSSIAN_LINSPACE_EPS, 1.0 - _GAUSSIAN_LINSPACE_EPS, size)
-        values = spec.mean + spec.std * math.sqrt(2.0) * torch.erfinv(2.0 * p - 1.0)
-        return values.to(dtype=dtype)
-
-    if dist == DistributionKind.LOG_UNIFORM_LINSPACE:
-        if spec.intervals:
-            for lo, hi in spec.intervals:
-                if lo <= 0 or hi <= 0:
-                    raise ValueError(
-                        f"log_uniform_linspace intervals require strictly positive "
-                        f"bounds, got ({lo}, {hi})"
-                    )
-            counts = _split_size_across_intervals(spec.intervals, size)
-            segments = []
-            for (lo, hi), n in zip(spec.intervals, counts):
-                if n > 0:
-                    log_lo = math.log(lo)
-                    log_hi = math.log(hi)
-                    segments.append(
-                        torch.exp(
-                            torch.linspace(log_lo, log_hi, n, dtype=torch.float32)
-                        )
-                    )
-            return torch.cat(segments).to(dtype=dtype)
-        if spec.low <= 0 or spec.high <= 0:
-            raise ValueError(
-                f"log_uniform_linspace requires strictly positive low and high, "
-                f"got low={spec.low}, high={spec.high}"
-            )
-        log_low = math.log(spec.low)
-        log_high = math.log(spec.high)
-        return torch.exp(
-            torch.linspace(log_low, log_high, size, dtype=torch.float32)
-        ).to(dtype=dtype)
-
-    raise ValueError(
-        f"_generate_linspace_tensor called with unsupported distribution {dist!r}"
-    )
-
-
-def _get_dtype_for_format(stimuli_format: DataFormat) -> torch.dtype:
-    """Return the torch dtype to use for *stimuli_format*."""
-    if stimuli_format in (DataFormat.Bfp8_b, DataFormat.Bfp4_b):
-        return torch.bfloat16
-    if stimuli_format == DataFormat.Tf32:
-        return torch.float32
-    return format_dict[stimuli_format]
-
-
-def _get_integer_bounds(stimuli_format: DataFormat) -> tuple[int, int]:
-    """Return the valid integer range (min, max) inclusive for *stimuli_format*.
-
-    For signed formats we exclude the most negative value (min + 1) because our
-    integer paths use sign-magnitude encoding, and the INT_MIN bit pattern cannot
-    be represented correctly in that scheme.
-    """
-    integer_format_dtype: Dict[DataFormat, torch.dtype] = {
-        DataFormat.Int8: torch.int8,
-        DataFormat.UInt8: torch.uint8,
-        DataFormat.Int16: torch.int16,
-        DataFormat.UInt16: torch.uint16,
-        DataFormat.Int32: torch.int32,
-        DataFormat.UInt32: torch.uint32,
-    }
-
-    if stimuli_format not in integer_format_dtype:
-        raise ValueError(f"Unsupported integer format: {stimuli_format}")
-
-    dtype = integer_format_dtype[stimuli_format]
-    info = torch.iinfo(dtype)
-    lo = info.min + 1 if info.min < 0 else info.min
-    return lo, info.max
-
-
-def _sample_uniform_intervals(
-    intervals: List[Tuple[float, float]],
-    size: int,
-    dtype: torch.dtype,
-    generator: Optional[torch.Generator],
-) -> torch.Tensor:
-    """Sample uniformly from a union of [low, high] intervals.
-
-    Each interval is selected with probability proportional to its length.
-    """
-    if not intervals:
-        raise ValueError("intervals must be a non-empty list")
-    lows = torch.tensor([lo for lo, _ in intervals], dtype=torch.float32)
-    highs = torch.tensor([hi for _, hi in intervals], dtype=torch.float32)
-    lengths = torch.clamp(highs - lows, min=0.0)
-    total = lengths.sum()
-    if total <= 0:
-        raise ValueError("Total interval length must be positive")
-    cdf = (lengths / total).cumsum(dim=0)
-    u = torch.rand(size, dtype=torch.float32, generator=generator)
-    idx = torch.searchsorted(cdf, u, right=False)
-    u_inner = torch.rand(size, dtype=torch.float32, generator=generator)
-    lo_sel = lows[idx]
-    hi_sel = highs[idx]
-    return (lo_sel + u_inner * (hi_sel - lo_sel)).to(dtype)
-
-
-def _sample_log_uniform_intervals(
-    intervals: List[Tuple[float, float]],
-    size: int,
-    dtype: torch.dtype,
-    generator: Optional[torch.Generator],
-) -> torch.Tensor:
-    """Sample log-uniformly from a union of strictly positive intervals.
-
-    Each interval is selected with probability proportional to its
-    log-space length (log(hi) - log(lo)).
-    """
-    if not intervals:
-        raise ValueError("intervals must be a non-empty list")
-    for lo, hi in intervals:
-        if lo <= 0 or hi <= 0:
-            raise ValueError(
-                f"log_uniform intervals require strictly positive bounds, "
-                f"got ({lo}, {hi})"
-            )
-    log_lows = torch.tensor([math.log(lo) for lo, _ in intervals], dtype=torch.float32)
-    log_highs = torch.tensor([math.log(hi) for _, hi in intervals], dtype=torch.float32)
-    lengths = torch.clamp(log_highs - log_lows, min=0.0)
-    total = lengths.sum()
-    if total <= 0:
-        raise ValueError("Total log-space interval length must be positive")
-    cdf = (lengths / total).cumsum(dim=0)
-    u = torch.rand(size, dtype=torch.float32, generator=generator)
-    idx = torch.searchsorted(cdf, u, right=False)
-    u_inner = torch.rand(size, dtype=torch.float32, generator=generator)
-    ll = log_lows[idx]
-    lh = log_highs[idx]
-    return torch.exp(ll + u_inner * (lh - ll)).to(dtype)
-
-
-def _sample_integer_intervals(
-    intervals: List[Tuple[float, float]],
-    int_min: int,
-    int_max: int,
-    size: int,
-    dtype: torch.dtype,
-    generator: Optional[torch.Generator],
-) -> torch.Tensor:
-    """Sample uniformly from a union of integer intervals (inclusive).
-
-    Each interval is selected with probability proportional to the
-    number of integer points it contains.
-    """
-    clamped = []
-    counts = []
-    for lo_f, hi_f in intervals:
-        lo_i = max(math.ceil(lo_f), int_min)
-        hi_i = min(math.floor(hi_f), int_max)
-        if hi_i >= lo_i:
-            clamped.append((lo_i, hi_i))
-            counts.append(hi_i - lo_i + 1)
-    if not clamped:
-        fallback = max(int_min, min((int_min + int_max) // 2, int_max))
-        warnings.warn(
-            f"No valid integer exists in any interval after clamping to "
-            f"[{int_min}, {int_max}]. Returning constant tensor filled with {fallback}.",
-            stacklevel=3,
-        )
-        return torch.full((size,), fallback, dtype=dtype)
-    lengths = torch.tensor(counts, dtype=torch.float32)
-    cdf = (lengths / lengths.sum()).cumsum(dim=0)
-    u = torch.rand(size, dtype=torch.float32, generator=generator)
-    idx = torch.searchsorted(cdf, u, right=False)
-    lows = torch.tensor([lo for lo, _ in clamped], dtype=torch.int64)
-    ranges = torch.tensor(counts, dtype=torch.int64)
-    lo_sel = lows[idx]
-    range_sel = ranges[idx]
-    u_inner = torch.rand(size, dtype=torch.float32, generator=generator)
-    offsets = (u_inner * range_sel.to(torch.float32)).to(torch.int64)
-    return (lo_sel + offsets).to(dtype)
-
-
-def _in_intervals(
-    values: torch.Tensor, intervals: List[Tuple[float, float]]
-) -> torch.Tensor:
-    """Return a boolean mask: True where the value falls inside any [lo, hi]."""
-    mask = torch.zeros(values.shape, dtype=torch.bool)
-    for lo, hi in intervals:
-        mask |= (values >= lo) & (values <= hi)
-    return mask
-
-
-def _split_size_across_intervals(
-    intervals: List[Tuple[float, float]], size: int
-) -> List[int]:
-    """Split *size* elements across intervals proportional to interval length.
-
-    Rounding residuals are distributed to the intervals with the largest
-    fractional parts so the total is exactly *size*.
-    """
-    if not intervals:
-        raise ValueError("intervals must be non-empty for piecewise linspace")
-    lengths = [abs(hi - lo) for lo, hi in intervals]
-    total_length = sum(lengths)
-    if total_length <= 0:
-        raise ValueError(
-            "Total interval length must be positive for piecewise linspace"
-        )
-    raw = [size * l / total_length for l in lengths]
-    counts = [int(r) for r in raw]
-    remainder = size - sum(counts)
-    fracs = sorted(range(len(counts)), key=lambda i: raw[i] - counts[i], reverse=True)
-    for j in range(remainder):
-        counts[fracs[j]] += 1
-    return counts
-
-
-def _sample_gaussian_intervals(
-    spec: StimuliSpec,
-    size: int,
-    dtype: torch.dtype,
-    generator: Optional[torch.Generator],
-) -> torch.Tensor:
-    """Sample from a truncated Gaussian over a union of intervals via rejection."""
-    _MAX_ATTEMPTS = 100
-    result = torch.empty(0, dtype=torch.float32)
-    remaining = size
-    for _ in range(_MAX_ATTEMPTS):
-        batch = max(remaining * 4, 64)
-        candidates = (
-            torch.randn(batch, dtype=torch.float32, generator=generator) * spec.std
-            + spec.mean
-        )
-        accepted = candidates[_in_intervals(candidates, spec.intervals)]
-        result = torch.cat([result, accepted])
-        if result.numel() >= size:
-            return result[:size].to(dtype)
-        remaining = size - result.numel()
-    raise ValueError(
-        f"Gaussian rejection sampling failed to fill {size} elements after "
-        f"{_MAX_ATTEMPTS} attempts. The intervals {spec.intervals} may be "
-        f"incompatible with mean={spec.mean}, std={spec.std}."
-    )
-
-
-def _sample_integer_gaussian_intervals(
-    spec: StimuliSpec,
-    int_min: int,
-    int_max: int,
-    size: int,
-    dtype: torch.dtype,
-    generator: Optional[torch.Generator],
-) -> torch.Tensor:
-    """Truncated integer Gaussian over a union of intervals via rejection.
-
-    When intervals are set, spec.low/spec.high are ignored — the domain is
-    spec.intervals intersected with the format's representable range.
-    """
-    clamped = []
-    for lo_f, hi_f in spec.intervals:
-        lo_i = max(math.ceil(lo_f), int_min)
-        hi_i = min(math.floor(hi_f), int_max)
-        if hi_i >= lo_i:
-            clamped.append((lo_i, hi_i))
-    if not clamped:
-        raise ValueError(
-            f"No valid integer interval remains after clamping "
-            f"{spec.intervals} to [{int_min}, {int_max}]."
-        )
-
-    def _in_int_intervals(vals: torch.Tensor) -> torch.Tensor:
-        mask = torch.zeros(vals.shape, dtype=torch.bool)
-        for lo_i, hi_i in clamped:
-            mask |= (vals >= lo_i) & (vals <= hi_i)
-        return mask
-
-    _MAX_ATTEMPTS = 100
-    result = torch.empty(0, dtype=torch.int64)
-    remaining = size
-    for _ in range(_MAX_ATTEMPTS):
-        batch = max(remaining * 4, 64)
-        raw = (
-            torch.randn(batch, dtype=torch.float32, generator=generator) * spec.std
-            + spec.mean
-        )
-        rounded = raw.round().clamp(int_min, int_max).to(torch.int64)
-        accepted = rounded[_in_int_intervals(rounded)]
-        result = torch.cat([result, accepted])
-        if result.numel() >= size:
-            return result[:size].to(dtype)
-        remaining = size - result.numel()
-    raise ValueError(
-        f"Integer Gaussian rejection sampling failed to fill {size} elements "
-        f"after {_MAX_ATTEMPTS} attempts. The intervals {spec.intervals} "
-        f"(clamped to {clamped}) may be incompatible with "
-        f"mean={spec.mean}, std={spec.std}."
-    )
-
-
-def _generate_integer_face(
-    spec: StimuliSpec,
-    stimuli_format: DataFormat,
-    size: int,
-    dtype: torch.dtype,
-    generator: Optional[torch.Generator],
-) -> torch.Tensor:
-    """Generate one face of integer-format data according to *spec*.
-
-    Bounds are narrowed to integers within [spec.low, spec.high] intersected
-    with the format's representable range.  Degenerate ranges produce constant
-    tensors (with a warning if no valid integer exists).
-
-    Args:
-        spec: Stimuli specification with distribution type and bounds.
-        stimuli_format: Target integer data format.
-        size: Number of elements per face.
-        dtype: Target torch dtype.
-        generator: Optional RNG state for reproducibility.
-
-    Returns:
-        1-D integer tensor of *size* elements.
-    """
-    int_min, int_max = _get_integer_bounds(stimuli_format)
-    # Narrow to the tightest set of integers that lie within [spec.low, spec.high]:
-    #   ceil(low)  = smallest integer ≥ spec.low
-    #   floor(high) = largest integer ≤ spec.high
-    # Then clamp to the format's representable range.
-    low = max(math.ceil(spec.low), int_min)
-    high = min(math.floor(spec.high), int_max)
-
-    if high < low:
-        # No integer exists in [spec.low, spec.high] after clamping —
-        # fall back to the nearest representable integer (midpoint, rounded).
-        fallback = max(int_min, min(round((spec.low + spec.high) / 2), int_max))
-        warnings.warn(
-            f"No integer exists in [{spec.low}, {spec.high}] for "
-            f"{stimuli_format.name} (representable range [{int_min}, {int_max}]). "
-            f"Returning constant tensor filled with {fallback}.",
-            stacklevel=3,
-        )
-        return torch.full((size,), fallback, dtype=dtype)
-
-    # Single valid integer — return constant tensor
-    if high == low:
-        return torch.full((size,), low, dtype=dtype)
-
-    distribution = spec.distribution
-
-    if distribution == DistributionKind.UNIFORM:
-        if spec.intervals:
-            return _sample_integer_intervals(
-                spec.intervals,
-                int_min,
-                int_max,
-                size,
-                dtype,
-                generator,
-            )
-        return torch.randint(
-            low=low, high=high + 1, size=(size,), dtype=dtype, generator=generator
-        )
-
-    if distribution == DistributionKind.SAW:
-        return (
-            torch.linspace(spec.low, spec.high, size, dtype=torch.float32)
-            .round()
-            .clamp(low, high)
-            .to(dtype=dtype)
-        )
-
-    if distribution == DistributionKind.GAUSSIAN:
-        if spec.intervals:
-            return _sample_integer_gaussian_intervals(
-                spec,
-                int_min,
-                int_max,
-                size,
-                dtype,
-                generator,
-            )
-        raw = (
-            torch.randn(size, dtype=torch.float32, generator=generator) * spec.std
-            + spec.mean
-        )
-        return raw.round().clamp(low, high).to(dtype=dtype)
-
-    if distribution == DistributionKind.LOG_UNIFORM:
-        if spec.low <= 0 or spec.high <= 0:
-            raise ValueError(
-                f"log_uniform requires strictly positive low and high bounds; "
-                f"got low={spec.low}, high={spec.high}.  "
-                f"For integer formats use a positive range such as "
-                f"StimuliSpec.log_uniform(low=1, high=1000)."
-            )
-        log_low = math.log(spec.low)
-        log_high = math.log(spec.high)
-        raw_u = torch.rand(size, dtype=torch.float32, generator=generator)
-        raw = torch.exp(raw_u * (log_high - log_low) + log_low)
-        return raw.round().clamp(low, high).to(dtype=dtype)
-
-    if distribution in _LINSPACE_DISTRIBUTIONS:
-        raw = _generate_linspace_tensor(spec, size, torch.float32)
-        return raw.round().clamp(low, high).to(dtype=dtype)
-
-    raise ValueError(
-        f"Unknown distribution {distribution!r} for integer format. "
-        f"Expected a DistributionKind member or a callable."
-    )
-
-
-def _generate_sequential(
-    spec: StimuliSpec,
-    size: int,
-    dtype: torch.dtype,
-    stimuli_format: Optional[DataFormat] = None,
-) -> torch.Tensor:
-    """Generate a sequential arithmetic progression.
-
-    Three modes:
-    1. *Default* (StimuliSpec.sequential() with no args): 1, 2, 3, …, size.
-    2. *Intervals* (spec.intervals is set): base sequence
-       1, 1+step, 1+2*step, … filling the full *size*, then values
-       outside the interval union are zeroed. spec.low/spec.high
-       are ignored.
-    3. *Custom* (low/high/step without intervals): start at
-       *low*, increment by *step*, zero-fill positions beyond *high*.
-    """
-    is_int = stimuli_format is not None and stimuli_format.is_integer()
-
-    # -- Mode 1: default (all dataclass defaults, no intervals) ---------------
-    if (
-        spec.low == 0.0
-        and spec.high == 1.0
-        and spec.std == 1.0
-        and spec.intervals is None
-    ):
-        if is_int:
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            result = torch.arange(1, size + 1, dtype=torch.int64)
-            return result.clamp(min=int_min, max=int_max).to(dtype)
-        return torch.arange(1, size + 1, dtype=dtype)
-
-    # -- Mode 2: intervals set — ignore low/high, use step + mask -----------
-    if spec.intervals is not None:
-        step = spec.std
-        idx = torch.arange(size, dtype=torch.float32)
-        vals = 1.0 + step * idx
-
-        if is_int:
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            vals = vals.round().clamp(int_min, int_max)
-
-        mask = _in_intervals(vals, spec.intervals)
-        vals = vals * mask.float()
-        return vals.to(dtype)
-
-    # -- Mode 3: custom low/high/step, no intervals -------------------------
-    low = spec.low
-    step = spec.std
-    high = spec.high
-
-    idx = torch.arange(size, dtype=torch.float32)
-    vals = low + step * idx
-
-    if high == 1.0 and low != 0.0:
-        # Special case: unbounded ramp — skip the zero-fill cutoff entirely.
-        pass
-    elif step > 0:
-        # Forward ramp: zero-fill positions where the value exceeds high
-        vals = torch.where(vals <= high, vals, torch.zeros_like(vals))
-    elif step < 0:
-        # Backward ramp: zero-fill positions where the value drops below high
-        vals = torch.where(vals >= high, vals, torch.zeros_like(vals))
-    # step == 0: every element equals low
-
-    if is_int:
-        int_min, int_max = _get_integer_bounds(stimuli_format)
-        vals = vals.round().clamp(int_min, int_max)
-
-    return vals.to(dtype)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Public: single-face generator
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def generate_face(
-    spec: StimuliSpec,
-    stimuli_format: DataFormat = DataFormat.Float16_b,
-    face_r_dim: int = MAX_FACE_R_DIM,
-    generator: Optional[torch.Generator] = None,
-) -> torch.Tensor:
-    """Generate a single face tensor of shape (face_r_dim * 16,) using *spec*.
-
-    Args:
-        spec: Generation specification (distribution, bounds, seed, etc.).
-        stimuli_format: Target hardware data format.
-        face_r_dim: Rows per face (1-16, default 16).
-        generator: External RNG state; when supplied, spec.seed is ignored.
-
-    Returns:
-        1-D tensor with face_r_dim * 16 elements.
-    """
-    size = face_r_dim * FACE_C_DIM
-    dtype = _get_dtype_for_format(stimuli_format)
-
-    # Create a local generator only when no external one was supplied
-    if generator is None and spec.seed is not None:
-        generator = torch.Generator()
-        generator.manual_seed(spec.seed)
-
-    distribution = spec.distribution
-
-    # ── custom callable ───────────────────────────────────────────────────────
-    if callable(distribution):
-        result = distribution(size, dtype, generator)
-        if not isinstance(result, torch.Tensor):
-            raise TypeError(
-                f"Custom distribution callable must return a torch.Tensor, "
-                f"got {type(result).__name__}"
-            )
-        if result.ndim != 1:
-            raise ValueError(
-                f"Custom distribution callable must return a 1-D tensor; "
-                f"got {result.ndim}-D tensor with shape {tuple(result.shape)}"
-            )
-        if len(result) != size:
-            raise ValueError(
-                f"Custom distribution callable returned {len(result)} elements "
-                f"but {size} were expected "
-                f"({face_r_dim} rows × {FACE_C_DIM} cols)"
-            )
-        return result.to(dtype=dtype)
-
-    # ── format-independent distributions ─────────────────────────────────────
-    if distribution == DistributionKind.CONSTANT:
-        if stimuli_format.is_integer():
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            clamped = max(int_min, min(int(spec.value), int_max))
-            return torch.full((size,), clamped, dtype=dtype)
-        return torch.full((size,), spec.value, dtype=dtype)
-
-    if distribution == DistributionKind.SEQUENTIAL:
-        return _generate_sequential(spec, size, dtype, stimuli_format)
-
-    if distribution == DistributionKind.IDENTITY:
-        raise ValueError(
-            "distribution='identity' is a tensor-level operation and cannot "
-            "be used in a per-face context (e.g. inside face_specs). "
-            "Use distribution='face_identity' for per-face identity blocks."
-        )
-
-    if distribution == DistributionKind.FACE_IDENTITY:
-        diag_val = spec.value
-        if stimuli_format.is_integer():
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            diag_val = max(int_min, min(int(round(diag_val)), int_max))
-        face = torch.zeros(face_r_dim, FACE_C_DIM, dtype=dtype)
-        face.diagonal()[:] = diag_val
-        return face.reshape(-1)
-
-    if distribution == DistributionKind.CUSTOM:
-        if spec.values is None or len(spec.values) == 0:
-            raise ValueError("distribution='custom' requires a non-empty 'values' list")
-        if len(spec.values) > size:
-            raise ValueError(
-                f"custom values list has {len(spec.values)} elements "
-                f"but face has only {size} "
-                f"({face_r_dim} rows × {FACE_C_DIM} cols)"
-            )
-        if stimuli_format.is_integer():
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            vals = [max(int_min, min(int(round(v)), int_max)) for v in spec.values]
-        else:
-            vals = list(spec.values)
-        tensor = torch.zeros(size, dtype=dtype)
-        tensor[: len(vals)] = torch.tensor(vals, dtype=dtype)
-        return tensor
-
-    # ── integer formats ───────────────────────────────────────────────────────
-    if stimuli_format.is_integer():
-        return _generate_integer_face(spec, stimuli_format, size, dtype, generator)
-
-    # ── float / BFP / MX formats ─────────────────────────────────────────────
-    if distribution == DistributionKind.UNIFORM:
-        if spec.intervals:
-            return _sample_uniform_intervals(spec.intervals, size, dtype, generator)
-        raw = torch.rand(size, dtype=dtype, generator=generator)
-        return raw * (spec.high - spec.low) + spec.low
-
-    if distribution == DistributionKind.GAUSSIAN:
-        if spec.intervals:
-            return _sample_gaussian_intervals(spec, size, dtype, generator)
-        raw = torch.randn(size, dtype=dtype, generator=generator)
-        return raw * spec.std + spec.mean
-
-    if distribution == DistributionKind.SAW:
-        if spec.intervals:
-            counts = _split_size_across_intervals(spec.intervals, size)
-            segments = []
-            for (lo, hi), n in zip(spec.intervals, counts):
-                if n > 0:
-                    segments.append(torch.linspace(lo, hi, n, dtype=dtype))
-            return torch.cat(segments)
-        return torch.linspace(spec.low, spec.high, size, dtype=dtype)
-
-    if distribution == DistributionKind.LOG_UNIFORM:
-        if spec.intervals:
-            return _sample_log_uniform_intervals(spec.intervals, size, dtype, generator)
-        if spec.low <= 0 or spec.high <= 0:
-            raise ValueError(
-                f"log_uniform requires strictly positive low and high, "
-                f"got low={spec.low}, high={spec.high}"
-            )
-        log_low = math.log(spec.low)
-        log_high = math.log(spec.high)
-        raw = torch.rand(size, dtype=dtype, generator=generator)
-        return torch.exp(raw * (log_high - log_low) + log_low).to(dtype=dtype)
-
-    if distribution in _LINSPACE_DISTRIBUTIONS:
-        return _generate_linspace_tensor(spec, size, dtype)
-
-    raise ValueError(
-        f"Unknown distribution {distribution!r}. "
-        f"Expected a DistributionKind member or a callable."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# ULP sweep helper
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def _enumerate_representable(
-    stimuli_format: DataFormat,
-    low: float,
-    high: float,
-    max_elements: int = 2**16,
-) -> torch.Tensor:
-    """Return all finite representable values in [low, high] for a 16-bit float format.
-
-    Iterates all 2^16 bit patterns, reinterprets them as the target dtype,
-    filters to finite values within [low, high], then sorts, deduplicates
-    (e.g. -0.0 == +0.0), and clips to max_elements.
-
-    Args:
-        stimuli_format: Must be Float16_b (bfloat16) or Float16 (float16).
-        low: Lower bound (inclusive).
-        high: Upper bound (inclusive).
-        max_elements: Safety cap on the number of returned values.
-
-    Returns:
-        1-D tensor in the native dtype of stimuli_format.
-
-    Raises:
-        ValueError: If stimuli_format is not Float16_b or Float16.
-    """
-    if stimuli_format == DataFormat.Float16_b:
-        dtype = torch.bfloat16
-    elif stimuli_format == DataFormat.Float16:
-        dtype = torch.float16
-    else:
-        raise ValueError(
-            f"ULP_SWEEP only supports Float16_b and Float16 formats, "
-            f"got {stimuli_format.name!r}"
-        )
-
-    # Enumerate all 2^16 bit patterns via int16 (wraps 32768..65535 to -32768..-1,
-    # but the byte representation is correct for a bitcast view).
-    all_bits = torch.arange(0, 2**16, dtype=torch.int16)
-    # Bitcast: both int16 and bfloat16/float16 are 2 bytes each.
-    all_vals = all_bits.view(dtype).to(torch.float32)
-
-    mask = torch.isfinite(all_vals) & (all_vals >= low) & (all_vals <= high)
-    vals = all_vals[mask]
-
-    vals, _ = torch.sort(vals)
-
-    if vals.numel() > 1:
-        unique_mask = torch.cat([torch.tensor([True]), vals[1:] != vals[:-1]])
-        vals = vals[unique_mask]
-
-    if vals.numel() > max_elements:
-        vals = vals[:max_elements]
-
-    return vals.to(dtype)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Private: full operand tensor
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def _generate_source_tensor(
-    stimuli_format: DataFormat,
-    num_elements: int,
-    face_r_dim: int,
-    spec: StimuliSpec,
-    input_dimensions: Optional[list] = None,
-) -> torch.Tensor:
-    """Generate a full operand tensor of *num_elements* using *spec*.
-
-    Iterates over faces with a shared RNG, applying face_specs overrides and
-    masked_faces zeroing.  Identity, sequential, and linspace distributions
-    short-circuit the face loop. BFP4_b tensors are post-quantized.
-
-    Args:
-        stimuli_format: Hardware data format.
-        num_elements: Total elements to generate.
-        face_r_dim: Rows per face (1-16).
-        spec: Stimuli specification.
-        input_dimensions: [rows, cols], required for identity distribution.
-
-    Returns:
-        1-D tensor of *num_elements* elements.
-    """
-    dtype = _get_dtype_for_format(stimuli_format)
-
-    # ── masked_faces validation for short-circuit distributions ──────────
-    _SHORT_CIRCUIT_DISTRIBUTIONS = frozenset(
-        {
-            DistributionKind.IDENTITY,
-            DistributionKind.SEQUENTIAL,
-            DistributionKind.ULP_SWEEP,
-        }
-        | _LINSPACE_DISTRIBUTIONS
-    )
-    if spec.masked_faces and spec.distribution in _SHORT_CIRCUIT_DISTRIBUTIONS:
-        raise ValueError(
-            f"masked_faces cannot be used with distribution={spec.distribution!r} "
-            f"because it bypasses the face loop.  Use a per-face distribution "
-            f"(uniform, gaussian, saw, …) instead."
-        )
-
-    # ── identity: tensor-level structured pattern (no face loop) ───────
-    if spec.distribution == DistributionKind.IDENTITY:
-        if input_dimensions is None or len(input_dimensions) != 2:
-            raise ValueError(
-                "distribution='identity' requires input_dimensions=[rows, cols]"
-            )
-        rows, cols = input_dimensions
-        diag_val = spec.value
-        if stimuli_format.is_integer():
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            diag_val = max(int_min, min(int(round(diag_val)), int_max))
-        tensor = torch.zeros(rows, cols, dtype=dtype)
-        tensor.diagonal()[:] = diag_val
-        tensor = tensor.reshape(-1)
-        if stimuli_format == DataFormat.Bfp4_b:
-            tensor = bfp4b_to_float16b(tensor)
-        return tensor
-
-    if spec.distribution == DistributionKind.SEQUENTIAL:
-        tensor = _generate_sequential(spec, num_elements, dtype, stimuli_format)
-        if stimuli_format == DataFormat.Bfp4_b:
-            tensor = bfp4b_to_float16b(tensor)
-        return tensor
-
-    # ── ULP sweep: enumerate all representable values (no face loop) ─────
-    if spec.distribution == DistributionKind.ULP_SWEEP:
-        vals = _enumerate_representable(stimuli_format, spec.low, spec.high)
-        n = vals.numel()
-        if n >= num_elements:
-            tensor = vals[:num_elements]
-        else:
-            padding = torch.zeros(num_elements - n, dtype=vals.dtype)
-            tensor = torch.cat([vals, padding])
-        if stimuli_format == DataFormat.Bfp4_b:
-            tensor = bfp4b_to_float16b(tensor)
-        return tensor
-
-    # ── linspace distributions: global sweep (no face loop) ──────────────
-    if spec.distribution in _LINSPACE_DISTRIBUTIONS:
-        if stimuli_format.is_integer():
-            int_min, int_max = _get_integer_bounds(stimuli_format)
-            tensor = _generate_linspace_tensor(spec, num_elements, torch.float32)
-            tensor = tensor.round().clamp(int_min, int_max).to(dtype=dtype)
-        else:
-            tensor = _generate_linspace_tensor(spec, num_elements, dtype)
-
-        if stimuli_format == DataFormat.Bfp4_b:
-            tensor = bfp4b_to_float16b(tensor)
-        return tensor
-
-    elements_per_face = face_r_dim * FACE_C_DIM
-    faces_needed = math.ceil(num_elements / elements_per_face)
-
-    gen: Optional[torch.Generator] = None
-    if spec.seed is not None:
-        gen = torch.Generator()
-        gen.manual_seed(spec.seed)
-
-    face_tensors: list[torch.Tensor] = []
-    for face_idx in range(faces_needed):
-        face_spec = spec
-        if spec.face_specs and face_idx < len(spec.face_specs):
-            override = spec.face_specs[face_idx]
-            if override is not None:
-                face_spec = override
-
-        face_tensor = generate_face(
-            spec=face_spec,
-            stimuli_format=stimuli_format,
-            face_r_dim=face_r_dim,
-            generator=gen,
-        )
-
-        # ── face masking: zero out selected faces ────────────────────────
-        if spec.masked_faces and face_idx in spec.masked_faces:
-            face_tensor = torch.zeros_like(face_tensor)
-
-        face_tensors.append(face_tensor)
-
-    tensor = torch.cat(face_tensors)[:num_elements]
-
-    # Simulate the hardware pack/unpack round-trip for BFP4_b so that stimuli
-    # already reflect the precision loss introduced by the format.
-    if stimuli_format == DataFormat.Bfp4_b:
-        tensor = bfp4b_to_float16b(tensor)
-
-    return tensor
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Built-in defaults for omitted specs
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def _default_bfp8b_face(
-    size: int, dtype: torch.dtype, gen: Optional[torch.Generator] = None
-) -> torch.Tensor:
-    integer_part = torch.randint(0, 3, (size,), generator=gen)
-    fraction = torch.randint(0, 16, (size,), generator=gen).to(torch.bfloat16) / 16.0
-    return integer_part.to(torch.bfloat16) + fraction
-
-
-def _default_bfp4b_face(
-    size: int, dtype: torch.dtype, gen: Optional[torch.Generator] = None
-) -> torch.Tensor:
-    integer_part = torch.randint(0, 3, (size,), generator=gen)
-    fraction = torch.randint(0, 8, (size,), generator=gen).to(torch.bfloat16) / 8.0
-    return integer_part.to(torch.bfloat16) + fraction
-
-
-def default_spec_for_format(stimuli_format: DataFormat) -> StimuliSpec:
-    """Return the built-in default StimuliSpec for a given data format.
-
-    Defaults are chosen to give reasonable value ranges and avoid overflows
-    (e.g. positive ranges for floats, half-range for integers).
-    """
-    if stimuli_format == DataFormat.MxFp8R:
-        return StimuliSpec.gaussian(
-            mean=0.1, std=0.05 * MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8R]
-        )
-    if stimuli_format == DataFormat.MxFp8P:
-        return StimuliSpec.gaussian(
-            mean=0.1, std=0.05 * MX_FORMAT_MAX_NORMAL[DataFormat.MxFp8P]
-        )
-    if stimuli_format == DataFormat.Bfp8_b:
-        return StimuliSpec(distribution=_default_bfp8b_face)
-    if stimuli_format == DataFormat.Bfp4_b:
-        return StimuliSpec(distribution=_default_bfp4b_face)
-    if stimuli_format.is_integer():
-        if stimuli_format == DataFormat.UInt32:
-            return StimuliSpec.uniform(low=0.0, high=float(2**32 - 2))
-        dtype = format_dict[stimuli_format]
-        v1_type_max = torch.iinfo(dtype).max // 2
-        return StimuliSpec.uniform(low=0.0, high=float(v1_type_max - 1))
-    return StimuliSpec.uniform(low=0.1, high=1.1)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Public: top-level stimuli generator
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def generate_stimuli(
-    stimuli_format_A: DataFormat = DataFormat.Float16_b,
-    input_dimensions_A: Optional[list] = None,
-    stimuli_format_B: DataFormat = DataFormat.Float16_b,
-    input_dimensions_B: Optional[list] = None,
-    spec_A: Optional[StimuliSpec] = None,
-    spec_B: Optional[StimuliSpec] = None,
-    tile_dimensions: Optional[list] = None,
-    face_r_dim: int = MAX_FACE_R_DIM,
-    num_faces: int = MAX_NUM_FACES,
-    output_format: Optional[DataFormat] = None,
-) -> tuple[torch.Tensor, int, torch.Tensor, int]:
-    """Generate test stimuli for two operands.
-
-    When tile_dimensions is provided, operates in dense mode with derived face layout;
-    otherwise uses the standard 32x32 tile path.
-
-    Args:
-        stimuli_format_A: Hardware data format for operand A.
-        input_dimensions_A: [height, width] in elements (default [32, 32]).
-        stimuli_format_B: Hardware data format for operand B.
-        input_dimensions_B: [height, width] in elements (default [32, 32]).
-        spec_A: Generation spec for operand A (default: format-aware built-in spec).
-        spec_B: Generation spec for operand B (default: format-aware built-in spec).
-        tile_dimensions: [rows, cols] tile size for dense mode (default None = standard path).
-        face_r_dim: Rows per face, 1-16 (ignored in dense mode, default 16).
-        num_faces: Faces per tile for partial-face case (ignored in dense mode, default 4).
-        output_format: Clamp outputs for mixed MX-format pairs when set.
-
-    Returns:
-        (srcA_tensor, tile_cnt_A, srcB_tensor, tile_cnt_B).
-    """
-    _spec_B_originally_none = spec_B is None
-
-    if input_dimensions_A is None:
-        input_dimensions_A = [DEFAULT_TILE_R_DIM, DEFAULT_TILE_C_DIM]
-    if input_dimensions_B is None:
-        input_dimensions_B = [DEFAULT_TILE_R_DIM, DEFAULT_TILE_C_DIM]
-    if spec_A is None:
-        spec_A = default_spec_for_format(stimuli_format_A)
-    if spec_B is None:
-        spec_B = default_spec_for_format(stimuli_format_B)
-
-    # ULP_SWEEP: auto-size input_dimensions_A from the count of representable values,
-    # ignoring any caller-supplied value. Mirror B if the caller didn't specify spec_B.
-    if spec_A.distribution == DistributionKind.ULP_SWEEP:
-        _ulp_vals = _enumerate_representable(stimuli_format_A, spec_A.low, spec_A.high)
-        _num_ulp = _ulp_vals.numel()
-        _tile_size = DEFAULT_TILE_R_DIM * DEFAULT_TILE_C_DIM
-        _num_tiles = max(1, math.ceil(_num_ulp / _tile_size))
-        # Align to 16 tiles when above the DestSync.Half limit (8) so the layout
-        # satisfies both DestSync.Half (needs num_tiles % 8 == 0) and DestSync.Full
-        # (needs num_tiles % 16 == 0 when > 16). LCM(8, 16) = 16.
-        if _num_tiles > 8:
-            _num_tiles = math.ceil(_num_tiles / 16) * 16
-        input_dimensions_A = [DEFAULT_TILE_R_DIM, DEFAULT_TILE_C_DIM * _num_tiles]
-        if _spec_B_originally_none:
-            input_dimensions_B = list(input_dimensions_A)
-
-    if tile_dimensions is not None:
-        if face_r_dim != MAX_FACE_R_DIM:
-            raise ValueError(
-                f"tile_dimensions and face_r_dim are mutually exclusive: "
-                f"when tile_dimensions is provided, face_r_dim is derived "
-                f"automatically. Got tile_dimensions={tile_dimensions}, "
-                f"face_r_dim={face_r_dim}."
-            )
-        if num_faces != MAX_NUM_FACES:
-            raise ValueError(
-                f"tile_dimensions and num_faces are mutually exclusive: "
-                f"when tile_dimensions is provided, num_faces is derived "
-                f"automatically.  Got tile_dimensions={tile_dimensions}, "
-                f"num_faces={num_faces}."
-            )
-        # Dense mode: derive face layout from tile_dimensions
-        face_r_dim, num_faces_r_dim, num_faces_c_dim = get_tile_params(tile_dimensions)
-        num_faces = num_faces_r_dim * num_faces_c_dim
-        tile_cnt_A, tile_cnt_B, _ = calculate_tile_and_face_counts_w_tile_dimensions(
-            input_dimensions_A,
-            input_dimensions_B,
-            face_r_dim,
-            num_faces,
-            tile_dimensions,
-        )
-    else:
-        # Standard 32×32 tile path
-        tile_cnt_A, tile_cnt_B, _ = calculate_tile_and_face_counts(
-            input_dimensions_A, input_dimensions_B, face_r_dim, num_faces
-        )
-
-    num_elements_A = input_dimensions_A[0] * input_dimensions_A[1]
-    num_elements_B = input_dimensions_B[0] * input_dimensions_B[1]
-
-    srcA_tensor = _generate_source_tensor(
-        stimuli_format=stimuli_format_A,
-        num_elements=num_elements_A,
-        face_r_dim=face_r_dim,
-        spec=spec_A,
-        input_dimensions=input_dimensions_A,
-    )
-    srcB_tensor = _generate_source_tensor(
-        stimuli_format=stimuli_format_B,
-        num_elements=num_elements_B,
-        face_r_dim=face_r_dim,
-        spec=spec_B,
-        input_dimensions=input_dimensions_B,
-    )
-
-    srcA_tensor, srcB_tensor = _clamp_mx_tensors(
-        srcA_tensor, srcB_tensor, stimuli_format_A, stimuli_format_B, output_format
-    )
-
-    return srcA_tensor, tile_cnt_A, srcB_tensor, tile_cnt_B
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Format-aware magnitude helpers
-# ─────────────────────────────────────────────────────────────────────────────
-
-
-def format_elem_max(data_format: DataFormat) -> float:
-    """
-    Return the maximum representable element magnitude for `data_format`.
-
-    Falls back to `torch.finfo(format_dict[data_format]).max` for non-MX formats.
-    """
-    if data_format.is_mx_format():
-        return MX_FORMAT_MAX_NORMAL[data_format]
-
-    return float(torch.finfo(format_dict[data_format]).max)
-
-
-def _format_elem_min_magnitude(data_format: DataFormat) -> float:
-    """
-    Return the minimum stimulus magnitude that avoids denormals / sub-range values
-    in `data_format`.
-
-    For MX formats uses the element type's documented minimum magnitude; for other
-    formats returns `max(1e-6, finfo(torch_dtype).tiny * 100)`.
-    """
-    if data_format.is_mx_format():
-        return MX_FORMAT_MIN_MAGNITUDE[data_format]
-
-    return max(1e-6, float(torch.finfo(format_dict[data_format]).tiny) * 100)
-
-
-def compute_safe_input_magnitude_range(
-    input_format: DataFormat,
-    output_format: DataFormat,
-    *,
-    input_magnitude_cap: float,
-    output_magnitude_cap: float,
-    bfloat16_precision_cap: float = 1e4,
-) -> tuple[float, float]:
-    """
-    Combine caller-supplied caps with format-aware rules to produce a
-    (min_magnitude, max_magnitude) range for stimuli feeding a tensix op.
-
-    The caller pre-computes both caps from the op's magnitude relation:
-      - `input_magnitude_cap`:  max |x| implied by the input format's range
-      - `output_magnitude_cap`: max |x| implied by the output format's range
-        (e.g. `sqrt(output_format_max)` for a squaring op, `output_format_max / 2`
-        for summing two operands, etc.)
-
-    The returned max magnitude is `min(input_magnitude_cap, output_magnitude_cap)`,
-    further clamped to `bfloat16_precision_cap` for non-MX bfloat16 inputs to keep
-    precision reasonable. Set `bfloat16_precision_cap=math.inf` to disable this.
-
-    The returned min magnitude respects MX minimum magnitudes when either input or
-    output is an MX format.
-    """
-    max_magnitude = min(input_magnitude_cap, output_magnitude_cap)
-
-    input_torch_format = format_dict[input_format]
-    if input_torch_format == torch.bfloat16 and not input_format.is_mx_format():
-        max_magnitude = min(max_magnitude, bfloat16_precision_cap)
-
-    min_magnitude = _format_elem_min_magnitude(input_format)
-    if output_format.is_mx_format():
-        min_magnitude = max(min_magnitude, _format_elem_min_magnitude(output_format))
-
-    return min_magnitude, max_magnitude
-
-
-def apply_log_uniform_magnitudes(
-    magnitude_source: torch.Tensor,
-    *,
-    min_magnitude: float,
-    max_magnitude: float,
-    cast_to_format: DataFormat,
-    sign_source: torch.Tensor = None,
-    alternate_sign_every_n: int = None,
-) -> torch.Tensor:
-    """
-    Remap the values in `magnitude_source` to log-uniform magnitudes in
-    [min_magnitude, max_magnitude] and apply signs, returning a tensor cast to
-    `format_dict[cast_to_format]`.
-
-    Sign selection (mutually exclusive):
-      - `sign_source` provided: normalized to [0, 1]; values < 0.5 map to -1, else +1
-      - `alternate_sign_every_n=n`: element i gets -1 when (i % n == 0), else +1
-      - neither provided: all +1
-
-    The output is clamped to [-max_magnitude, max_magnitude] before casting.
-    """
-    if sign_source is not None and alternate_sign_every_n is not None:
-        raise ValueError(
-            "Provide either sign_source or alternate_sign_every_n, not both"
-        )
-
-    src_float = magnitude_source.to(torch.float32)
-    src_min = src_float.min()
-    src_max = src_float.max()
-    normalized = (
-        (src_float - src_min) / (src_max - src_min)
-        if src_max > src_min
-        else torch.zeros_like(src_float)
-    )
-
-    log_min = torch.log(torch.tensor(min_magnitude, dtype=torch.float32))
-    log_max = torch.log(torch.tensor(max_magnitude, dtype=torch.float32))
-    magnitudes = torch.exp(log_min + normalized * (log_max - log_min))
-
-    if sign_source is not None:
-        sign_float = sign_source.to(torch.float32)
-        sign_min = sign_float.min()
-        sign_max = sign_float.max()
-        sign_normalized = (
-            (sign_float - sign_min) / (sign_max - sign_min)
-            if sign_max > sign_min
-            else torch.zeros_like(sign_float)
-        )
-        signs = torch.where(sign_normalized < 0.5, -1.0, 1.0)
-    elif alternate_sign_every_n is not None:
-        signs = torch.where(
-            torch.arange(magnitude_source.numel()) % alternate_sign_every_n == 0,
-            torch.tensor(-1.0),
-            torch.tensor(1.0),
-        )
-    else:
-        signs = torch.ones_like(magnitudes)
-
-    values = signs * magnitudes
-    values = torch.clamp(values, -max_magnitude, max_magnitude)
-    return values.to(format_dict[cast_to_format])
